requirements.txt
Flask
Flask_SQLAlchemy
Flask-Migrate
python-dotenv
requests
pymysql
pytest
